+++ src/main/java/com/sparta/memo/controller/Memo3RAController.java
@@ -1,0 +1,44 @@
+package com.sparta.memo.controller;
+
+import com.sparta.memo.dto.MemoRequestDto;
+import com.sparta.memo.dto.MemoResponseDto;
+import com.sparta.memo.service.MemoService;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api")
+public class Memo3RAController {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public Memo3RAController(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @PostMapping("/memos")
+    public MemoResponseDto createMemo(@RequestBody MemoRequestDto requestDto) {
+        MemoService memoService = new MemoService(jdbcTemplate);
+        return memoService.createMemo(requestDto);
+    }
+
+    @GetMapping("/memos")
+    public List<MemoResponseDto> getMemos() {
+        MemoService memoService = new MemoService(jdbcTemplate);
+        return memoService.getMemos();
+    }
+
+    @PutMapping("/memos/{id}")
+    public Long updateMemo(@PathVariable Long id, @RequestBody MemoRequestDto requestDto) {
+        MemoService memoService = new MemoService(jdbcTemplate);
+        return memoService.updateMemo(id, requestDto);
+    }
+
+    @DeleteMapping("/memos/{id}")
+    public Long deleteMemo(@PathVariable Long id) {
+        MemoService memoService = new MemoService(jdbcTemplate);
+        return memoService.deleteMemo(id);
+    }
+}
